make ChainTokens single component to display balances on hover over chain icon
