+++ packages/synapse-interface/components/Activity/AirdropRewards.tsx
@@ -5,7 +5,7 @@ import numeral from 'numeral'
 import { Address, useAccount } from 'wagmi'
 import { arbitrum } from 'viem/chains'
 import { useAppSelector } from '@/store/hooks'
-import { useState, useEffect, useRef } from 'react'
+import { useState, useEffect } from 'react'
 import { trimTrailingZeroesAfterDecimal } from '@/utils/trimTrailingZeroesAfterDecimal'
 import { getErc20TokenTransfers } from '@/utils/actions/getErc20TokenTransfers'
 import { formatBigIntToString } from '@/utils/bigint/format'
@@ -14,7 +14,7 @@ import { ARBITRUM } from '@/constants/chains/master'
 import { CloseButton } from '@/components/buttons/CloseButton'
 import { ArrowUpRightIcon } from '../icons/ArrowUpRightIcon'
 import { QuestionMarkCircleIcon } from '@heroicons/react/outline'
-import { HoverContent } from '@/components/Portfolio/components/HoverTokenAndBalance'
+import { HoverContent } from '@/components/Portfolio/components/HoverContent'
 
 import TransactionArrow from '../icons/TransactionArrow'
 import arbitrumImg from '@assets/chains/arbitrum.svg'
