+++ packages/synapse-interface/components/Portfolio/components/ChainTokens.tsx
@@ -1,32 +1,33 @@
-import { useState } from 'react'
-import Image from 'next/image'
-import { TokenAndBalance } from '@/utils/actions/fetchPortfolioBalances'
-import { HoverTokenAndBalance } from './HoverTokenAndBalance'
+import { useState } from "react"
+import Image from "next/image"
 
-export const PortfolioTokenVisualizer = ({
-  portfolioTokens,
+import { TokenAndBalance } from "@/utils/actions/fetchPortfolioBalances"
+import { HoverTokenAndBalance } from "./HoverTokenAndBalance"
+
+export const ChainTokens = ({
+  balanceTokens=[],
+  hoverClassName="",
 }: {
-  portfolioTokens: TokenAndBalance[]
+  balanceTokens?: TokenAndBalance[]
+  hoverClassName?: string
 }) => {
   const [isT1Hovered, setIsT1Hovered] = useState<boolean>(false)
   const [isT2Hovered, setIsT2Hovered] = useState<boolean>(false)
   const [isT3Hovered, setIsT3Hovered] = useState<boolean>(false)
 
   const hasNoTokens: boolean =
-    !portfolioTokens || (portfolioTokens && portfolioTokens.length === 0)
-  const hasOneToken: boolean = portfolioTokens && portfolioTokens.length > 0
-  const hasTwoTokens: boolean = portfolioTokens && portfolioTokens.length > 1
+    !balanceTokens || (balanceTokens?.length === 0)
+  const hasOneToken: boolean = balanceTokens?.length > 0
+  const hasTwoTokens: boolean = balanceTokens?.length > 1
   const numOverTwoTokens: number =
-    portfolioTokens && portfolioTokens.length - 2 > 0
-      ? portfolioTokens.length - 2
-      : 0
-  const hasOnlyOneToken: boolean =
-    portfolioTokens && portfolioTokens.length === 1
+    balanceTokens?.length - 2 > 0 ? balanceTokens.length - 2 : 0
+  const hasOnlyOneToken: boolean = balanceTokens?.length === 1
+
 
   if (hasNoTokens) {
     return (
       <div
-        id="portfolio-token-visualizer"
+        data-test-id="portfolio-token-visualizer"
         className="flex flex-row items-center mr-4 cursor-pointer hover-trigger text-secondary"
       >
         -
@@ -35,7 +36,7 @@ export const PortfolioTokenVisualizer = ({
   }
   return (
     <div
-      id="portfolio-token-visualizer"
+      data-test-id="portfolio-token-visualizer"
       className="flex flex-row items-center space-x-2 cursor-pointer hover-trigger"
     >
       {hasOneToken && (
@@ -43,20 +44,21 @@ export const PortfolioTokenVisualizer = ({
           <Image
             loading="lazy"
             className="w-6 h-6 rounded-md"
-            alt={`${portfolioTokens[0].token.symbol} img`}
-            src={portfolioTokens[0].token.icon}
+            alt={`${balanceTokens[0].token.symbol} img`}
+            src={balanceTokens[0].token.icon}
             onMouseEnter={() => setIsT1Hovered(true)}
             onMouseLeave={() => setIsT1Hovered(false)}
           />
           <HoverTokenAndBalance
             isHovered={isT1Hovered}
-            tokens={[portfolioTokens[0]]}
+            tokens={[balanceTokens[0]]}
+            hoverClassName={hoverClassName}
           />
         </div>
       )}
       {hasOnlyOneToken && (
         <div className="text-white whitespace-nowrap">
-          {portfolioTokens[0].parsedBalance} {portfolioTokens[0].token.symbol}
+          {balanceTokens[0].parsedBalance} {balanceTokens[0].token.symbol}
         </div>
       )}
       {hasTwoTokens && (
@@ -64,14 +66,15 @@ export const PortfolioTokenVisualizer = ({
           <Image
             loading="lazy"
             className="w-6 h-6 rounded-md"
-            alt={`${portfolioTokens[1].token.symbol} img`}
-            src={portfolioTokens[1].token.icon}
+            alt={`${balanceTokens[1].token.symbol} img`}
+            src={balanceTokens[1].token.icon}
             onMouseEnter={() => setIsT2Hovered(true)}
             onMouseLeave={() => setIsT2Hovered(false)}
           />
           <HoverTokenAndBalance
             isHovered={isT2Hovered}
-            tokens={[portfolioTokens[1]]}
+            tokens={[balanceTokens[1]]}
+            hoverClassName={hoverClassName}
           />
         </div>
       )}
@@ -86,10 +89,10 @@ export const PortfolioTokenVisualizer = ({
       )}
       <HoverTokenAndBalance
         isHovered={isT3Hovered}
-        tokens={portfolioTokens}
+        tokens={balanceTokens}
         startFrom={2}
+        hoverClassName={`${hoverClassName} mt-3`}
       />
     </div>
   )
 }
-
