+++ packages/synapse-interface/components/StateManagedBridge/components/SelectSpecificNetworkButton.tsx
@@ -21,7 +21,7 @@ import {
   useStipEligibility,
 } from '@/utils/hooks/useStipEligibility'
 import { useBridgeState } from '@/slices/bridge/hooks'
-import { HoverTokenAndBalance } from '@/components/Portfolio/components/HoverTokenAndBalance'
+import { ChainTokens } from '@/components/Portfolio/components/ChainTokens'
 
 export const SelectSpecificNetworkButton = ({
   itemChainId,
@@ -126,99 +126,11 @@ function ButtonContent({
         </div>
       </div>
       {isOrigin && balanceTokens && balanceTokens.length > 0 ? (
-        <ChainTokens balanceTokens={balanceTokens} />
+        <ChainTokens
+          balanceTokens={balanceTokens}
+          hoverClassName="bg-slate-900/70"
+        />
       ) : null}
     </>
   ) : null
-}
-
-const ChainTokens = ({
-  balanceTokens = [],
-}: {
-  balanceTokens: TokenAndBalance[]
-}) => {
-  const [isT1Hovered, setIsT1Hovered] = useState<boolean>(false)
-  const [isT2Hovered, setIsT2Hovered] = useState<boolean>(false)
-  const [isT3Hovered, setIsT3Hovered] = useState<boolean>(false)
-
-  const hasNoTokens: boolean =
-    !balanceTokens || (balanceTokens && balanceTokens.length === 0)
-  const hasOneToken: boolean = balanceTokens && balanceTokens.length > 0
-  const hasTwoTokens: boolean = balanceTokens && balanceTokens.length > 1
-  const numOverTwoTokens: number =
-    balanceTokens && balanceTokens.length - 2 > 0 ? balanceTokens.length - 2 : 0
-  const hasOnlyOneToken: boolean = balanceTokens && balanceTokens.length === 1
-  const hasOnlyTwoTokens: boolean = balanceTokens && balanceTokens.length === 2
-
-  if (hasNoTokens) {
-    return (
-      <div
-        data-test-id="portfolio-token-visualizer"
-        className="flex flex-row items-center mr-4 cursor-pointer hover-trigger text-secondary"
-      >
-        -
-      </div>
-    )
-  }
-  return (
-    <div
-      data-test-id="portfolio-token-visualizer"
-      className="flex flex-row items-center space-x-2 cursor-pointer hover-trigger"
-    >
-      {hasOneToken && (
-        <div>
-          <Image
-            loading="lazy"
-            className="w-6 h-6 rounded-md"
-            alt={`${balanceTokens[0].token.symbol} img`}
-            src={balanceTokens[0].token.icon}
-            onMouseEnter={() => setIsT1Hovered(true)}
-            onMouseLeave={() => setIsT1Hovered(false)}
-          />
-          <HoverTokenAndBalance
-            isHovered={isT1Hovered}
-            tokens={[balanceTokens[0]]}
-            hoverClassName="bg-slate-900/70"
-          />
-        </div>
-      )}
-      {hasOnlyOneToken && (
-        <div className="text-white whitespace-nowrap">
-          {balanceTokens[0].parsedBalance} {balanceTokens[0].token.symbol}
-        </div>
-      )}
-      {hasTwoTokens && (
-        <div>
-          <Image
-            loading="lazy"
-            className="w-6 h-6 rounded-md"
-            alt={`${balanceTokens[1].token.symbol} img`}
-            src={balanceTokens[1].token.icon}
-            onMouseEnter={() => setIsT2Hovered(true)}
-            onMouseLeave={() => setIsT2Hovered(false)}
-          />
-          <HoverTokenAndBalance
-            isHovered={isT2Hovered}
-            tokens={[balanceTokens[1]]}
-            hoverClassName="bg-slate-900/70"
-          />
-        </div>
-      )}
-      {numOverTwoTokens > 0 && (
-        <div
-          className="text-white"
-          onMouseEnter={() => setIsT3Hovered(true)}
-          onMouseLeave={() => setIsT3Hovered(false)}
-        >
-          + {numOverTwoTokens}
-        </div>
-      )}
-      <HoverTokenAndBalance
-        isHovered={isT3Hovered}
-        tokens={balanceTokens}
-        startFrom={2}
-        hoverClassName="bg-slate-900/70"
-      />
-    </div>
-  )
 }
